discovery/multi: move SignAnnouncement to netann
